move util function to util module
